# docs/swagger.yaml
basePath: /api/
consumes:
- application/json
definitions:
  CancelSaleSuccessfulResposnse:
    properties:
      mesage:
        example: sale cancelled successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  GetAllSalesSuccessfulResponse:
    properties:
      data:
        $ref: '#/definitions/sale.GetAllSales'
      mesage:
        example: fetched all current sales
        type: string
      success:
        example: true
        type: boolean
    type: object
  GetPersonalBidsSuccessfulResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Bid'
        type: array
      mesage:
        example: fetched all sale bids
        type: string
      success:
        example: true
        type: boolean
    type: object
  GetPersonalSalesSuccessfulResponse:
    properties:
      data:
        $ref: '#/definitions/sale.GetAllSales'
      mesage:
        example: fetched all sale bids
        type: string
      success:
        example: true
        type: boolean
    type: object
  GetSaleBidsSuccessfulResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Bid'
        type: array
      mesage:
        example: fetched all sale bids
        type: string
      success:
        example: true
        type: boolean
    type: object
  GetSaleSuccessfulResponse:
    properties:
      data:
        $ref: '#/definitions/models.Sale'
      mesage:
        example: fetched all current sales
        type: string
      success:
        example: true
        type: boolean
    type: object
  Login:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  Register:
    properties:
      accountType:
        $ref: '#/definitions/models.AccountType'
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - accountType
    - email
    - password
    - username
    type: object
  auth.LoginSuccessfulResponse:
    properties:
      data:
        $ref: '#/definitions/auth.UserAuthData'
      mesage:
        example: account created successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  auth.RegistrationSuccessfulResponse:
    properties:
      data:
        $ref: '#/definitions/auth.UserAuthData'
      mesage:
        example: account created successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  auth.UserAuthData:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.AccountType:
    enum:
    - BUYER
    - SELLER
    type: string
    x-enum-varnames:
    - BUYER
    - SELLER
  models.Bid:
    properties:
      amount:
        type: number
      createdAt:
        example: 2022-08-21 21:08
        readOnly: true
        type: string
      id:
        readOnly: true
        type: string
      saleId:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  models.Sale:
    properties:
      bidCount:
        type: integer
      bids:
        items:
          $ref: '#/definitions/models.Bid'
        type: array
      breed:
        type: string
      category:
        example: DOG
        type: string
      createdAt:
        example: 2022-08-21 21:08
        readOnly: true
        type: string
      description:
        type: string
      expiresBy:
        type: string
      id:
        readOnly: true
        type: string
      priority:
        $ref: '#/definitions/models.SalePriority'
      sold:
        type: boolean
      soldTo:
        type: string
      startingBig:
        type: number
      status:
        $ref: '#/definitions/models.SaleStatus'
      title:
        type: string
      traderId:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/models.SaleType'
        example: type
      updatedAt:
        type: string
    type: object
  models.SalePriority:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - LOW
    - NORMAL
    - HIGH
  models.SaleStatus:
    enum:
    - PENDING
    - PUBLISHED
    - CANCELLED
    - CLOSED
    type: string
    x-enum-varnames:
    - PENDING
    - PUBLISHED
    - CANCELLED
    - CLOSED
  models.SaleType:
    enum:
    - NEW
    - PROMOTED
    - REPUBLISHED
    type: string
    x-enum-varnames:
    - NEW
    - PROMOTED
    - REPUBLISHED
  models.User:
    properties:
      accountType:
        $ref: '#/definitions/models.AccountType'
      bids:
        items:
          $ref: '#/definitions/models.Bid'
        type: array
      createdAt:
        example: 2022-08-21 21:08
        readOnly: true
        type: string
      email:
        type: string
      id:
        readOnly: true
        type: string
      password:
        type: string
      profileImg:
        type: string
      sales:
        items:
          $ref: '#/definitions/models.Sale'
        type: array
      updatedAt:
        type: string
      username:
        type: string
      wallet:
        $ref: '#/definitions/models.Wallet'
    type: object
  models.Wallet:
    properties:
      balance:
        type: number
      createdAt:
        example: 2022-08-21 21:08
        readOnly: true
        type: string
      id:
        readOnly: true
        type: string
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  sale.CreateBidSuccessfulResponse:
    properties:
      data:
        $ref: '#/definitions/models.Sale'
      mesage:
        example: bid placed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  sale.CreateSaleSuccessfulResponse:
    properties:
      data:
        additionalProperties:
          type: string
        example:
          bidCount: "0"
          breed: Cat Dog
          category: CAT
          createdAt: "2023-01-14T02:06:31+01:00"
          description: BDG for sale
          id: clcv8zrk90000vboy7d4tb5ms
          priority: "1"
          sold: "false"
          startingBig: "20000"
          status: PENDING
          title: Bull Dog For Sale
          traderId: clcnz7h460000s1mo30jmkewi
          type: NEW
          updatedAt: "2023-01-14T02:06:31+01:00"
        type: object
      mesage:
        example: sale created successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  sale.GetAllSales:
    properties:
      meta:
        $ref: '#/definitions/utils.Pagination'
      sales:
        items:
          $ref: '#/definitions/models.Sale'
        type: array
    type: object
  user.GetProfileSuccessfulResponse:
    properties:
      data:
        $ref: '#/definitions/models.User'
      mesage:
        example: fetched user profile
        type: string
      success:
        example: true
        type: boolean
    type: object
  utils.Pagination:
    properties:
      nextPage:
        type: integer
      previousPage:
        type: integer
      size:
        type: integer
      totalItems:
        type: integer
    type: object
host: localhost:3000
info:
  contact: {}
  description: Pawdow API Service
  license:
    name: ISC
  title: Pawdot API Service
  version: "1.0"
paths:
  /auth/login:
    post:
      operationId: login
      parameters:
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Login'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.LoginSuccessfulResponse'
      tags:
      - Auth
  /auth/register:
    post:
      operationId: register
      parameters:
      - description: Body
        in: body
        name: Register
        required: true
        schema:
          $ref: '#/definitions/Register'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.RegistrationSuccessfulResponse'
      tags:
      - Auth
  /marketplace/personal/bids:
    get:
      operationId: getPersonalBids
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/GetPersonalBidsSuccessfulResponse'
      security:
      - ApiKey: []
      tags:
      - Bid
  /marketplace/personal/sales:
    get:
      operationId: getPersonalSales
      parameters:
      - description: Breed
        in: query
        name: breed
        type: string
      - description: Category
        in: query
        name: category
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/GetPersonalSalesSuccessfulResponse'
      security:
      - ApiKey: []
      tags:
      - Sale
  /marketplace/sales:
    get:
      operationId: getAllSales
      parameters:
      - description: Breed
        in: query
        name: breed
        type: string
      - description: Category
        in: query
        name: category
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/GetAllSalesSuccessfulResponse'
      security:
      - ApiKey: []
      tags:
      - Sale
  /marketplace/sales/:saleID:
    get:
      operationId: getSale
      parameters:
      - description: saleID
        in: path
        name: saleID
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/GetSaleSuccessfulResponse'
      security:
      - ApiKey: []
      tags:
      - Sale
  /marketplace/sales/:saleID/bid:
    post:
      operationId: creaateBid
      parameters:
      - description: saleID
        in: path
        name: saleID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sale.CreateBidSuccessfulResponse'
      security:
      - ApiKey: []
      tags:
      - Bid
  /marketplace/sales/:saleID/bids:
    get:
      operationId: getSaleBids
      parameters:
      - description: saleID
        in: path
        name: saleID
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/GetSaleBidsSuccessfulResponse'
      security:
      - ApiKey: []
      tags:
      - Bid
  /marketplace/sales/:saleID/cancel:
    delete:
      operationId: cancelSale
      parameters:
      - description: saleID
        in: path
        name: saleID
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CancelSaleSuccessfulResposnse'
      security:
      - ApiKey: []
      tags:
      - Sale
  /marketplace/sales/:saleID/republish:
    post:
      operationId: republishSale
      parameters:
      - description: saleID
        in: path
        name: saleID
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CancelSaleSuccessfulResposnse'
      security:
      - ApiKey: []
      tags:
      - Sale
  /marketplace/sales/new:
    post:
      operationId: createSale
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sale.CreateSaleSuccessfulResponse'
      security:
      - ApiKey: []
      tags:
      - Sale
  /user/profile:
    get:
      operationId: profile
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.GetProfileSuccessfulResponse'
      security:
      - ApiKey: []
      tags:
      - User
produces:
- application/json
securityDefinitions:
  ApiKey:
    in: header
    name: x-auth-token
    type: apiKey
swagger: "2.0"